organize makefile

+++ AppMatch/src/base/CMatchListValid.cpp
@@ -63,8 +63,8 @@ bool NSMatch::isListValid_count(const MTermSet & Comb)
     {
         int cc_open = 0;
         int cc_close = 0;
-        int ns = (Comb.size());
-        for (int j = 0; j < ns; ++j) {
+        size_t  ns = (Comb.size());
+        for (size_t j = 0; j < ns; ++j) {
             // if ((*it).at(j)->repr() == ")") cc_open++;
             // if ((*it).at(j)->repr() == "(") cc_close++;
